Compiles cleanly without MPI Original: 40a61fef7903953959f4cd10798e65993717020a

+++ lalinference/src/LALInferenceTemplate.c
@@ -1622,7 +1622,7 @@ void templateLALGenerateInspiral(LALIFOData *IFOdata)
           }
 				}else if(waveform.h){
           for (i=0; i<IFOdata->timeData->data->length; i++){
-            if((i+1)>=((int)(waveform.h->data->length) - 1)){
+            if((i+1)>=((unsigned long int)(waveform.h->data->length) - 1)){
               IFOdata->timeModelhPlus->data->data[i] = 0;
               IFOdata->timeModelhCross->data->data[i] = 0;		
             }else{
